+++ et-common-libs/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Optimized et_printf by removing extra vsnprintf call.
 - [SW-10869] replaced cache inline functions with macros
+- Require esperantoTrace/0.6.0
 ### Deprecated
 ### Removed
 ### Fixed