+++ et-common-libs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [SW-13832] Add SP2MM GET_MM_STATS command and response. Also fixed some typos.
+- [SW-13832] Update SP2MM GET_MM_STATS command response to include a status field.
 ### Changed
 ### Deprecated
 ### Removed